+++ sso/api/views_activity_stream.py
@@ -97,6 +97,9 @@ class _ActivityStreamHawkResponseMiddleware:
     of the request can authenticate the response
     """
 
+    def __init__(self, *args, **kwargs):
+        ...
+
     def process_response(self, viewset, response):
         """Adds the Server-Authorization header to the response, so the originator
         of the request can authenticate the response