+++ CHANGELOG.md
@@ -8,8 +8,11 @@
 
 ### Enhancements
 
+- GLS-380 - Upgrade package to use Django 3.2
+
 ## [9.6.0](https://github.com/uktrade/directory-sso/releases/tag/9.6.0)
 [Compare](https://github.com/uktrade/directory-sso/compare/9.5.0...9.6.0)
+
 ### Enhancements
 
 - GLS-336 - Resend verification code if expired code is entered